Resolve Demon/Undead confusion
Undead and Demon were misordered in one version of the comments which happens to be the one I took the enum titles from. I ended up looking at the scenario files in the scenario editor so that I could finally see the correct order and it was the other way around.
Note that the compilation before and after this rename generated exacty the same executable/data which validates I haven't changed anything.