Refactoring 8 (#54)
* Use reference to pc rather than array + offset so that spinning out to separate function diff  is simpler

* Small no effect tweak to separate out pc specific calculations from party ones

* Pull parry modifer outside of general logic in preparation for extraction

* Pull party adjustment outside of general logic in preparation for extraction

* Extract pc_damage_adjust  logic into own function

* Tidy up call site to pc_damage_adjust

* Move pc_damage_adjust to new home

* Reformat pc_damage_adjust

* More pc_damage_adjust tidy up

* Resolve Demon/Undead confusion

Undead and Demon were misordered in one version of the comments which happens to be the one I took the enum titles from. I ended up looking at the scenario files in the scenario editor so that I could finally see the correct order and it was the other way around.
Note that the compilation before and after this rename generated exacty the same executable/data which validates I haven't changed anything.

* Try to improve type safety for const functions and stop overloading the damage_type to carry extra info

This step uses some temporary consteval functions to compile time calculate the actual parameter values which would have been passed.
The next step will be to change these to the values I *think* they should be and check the compiled binary looks identical.

* Remove sticking plaster

Before/after executables show no change to code or data, so we're good.

* Try to improve type safety for const functions and stop overloading the damage_type to carry extra info

This step uses some temporary consteval functions to compile time calculate the actual parameter values which would have been passed. The next step will be to change these to the values I *think* they should be and check the compiled binary looks identical.

* Remove sticking plaster

Before/after executables show no change to code or data, so we're good.

* Rename stat_adj to pc_stat_adj ahead of move to pc.cpp

* Factor our simple skill_bonus function

* Move skill bonus to new home

* Move pc_stat_adj to common library

* Extract pc_poison function

* Change signature to take pc_record_type and no longer rely on global adven

* Move pc_poison to common library

* Tidy up

* Invert check to make next change easier

* Leverage has_trait member function to make code more readable

The issue is that the check for the traits is done in three different ways and unfortunately since the real data is a "short" rather than a bool this will need validation (and possibly correction) at party load time before it can be proven to only need one variant.

* A few more trait checks switched to the member function

Proven no change by comparing code & data sections of compiled exe

* Move some spell details into common library

* Re-org and move to std::array

* Provide mage_spell_cast and priest_spell_cast

Initially in the form of a macro just to prove before/after produce an identical binary

* Change macro's to functions now we have proven there is no effect

* Resolve regression where PC Editor and Scenario text doesn't appear in some dialogs

See https://github.com/CorrodedCoder/Blades-of-Exile/issues/53 for the ugly details.