Change remaining `long` game types to `int`
The game was written assuming that long was 32-bit, but on 64-bit platforms it is not always so (depending on whether it's an LP64 compiler or just a P64 compiler like MSVC).
The remaining untouched longs are actually being used to track time duration so they will eventually change to a more suitable type.