Move resources to new home (#66)
* Move bitmap files out of source tree and into resource area

* Copy files from the Blades of Exile installer

This works around some of the corruption/missing files detailed in:
https://github.com/CorrodedCoder/Blades-of-Exile/issues/63 and
https://github.com/CorrodedCoder/Blades-of-Exile/issues/64

Some files come from the demo and some from the full version.

* Add in files from full game install

This additional files were not included in the original source download, but were part of the demo and/or full versions of the game also provided for free.

* Resolve https://github.com/CorrodedCoder/Blades-of-Exile/issues/65

Revert back to original behaviour and *be* *more* *careful* in future!

* Correct path to scenarios

* Only copy expected files to output areas

Along with bmp files, also copy .hlp files, bladbase.exs and also bladdata.bld to effectively make it appear as a registered game by default. Also change the animation from the original/demo version to the registered one for now.

* Add original look.cur into the project

This file was extracted from the officially released registered executable.