Extract LoadDibData helper
As part of this change I discovered that GetOpenFileName and GetSaveFileName actually change the current directory.
I guess this explains the mystery of why there was this complicated logic for looking up the directory whenever a data file was being loaded.
Anyway, I was able to change the Flags part of the relevant structures to include OFN_NOCHANGEDIR which means the directory is no longer changed and therefore the loading logic elsewhere becomes a straightforward relative path.